requirements.txt
# Core Machine Learning
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.23.0

# Data Acquisition & Preprocessing
tensorflow-datasets>=4.8.0
Pillow>=9.4.0

# Visualization & Analysis
matplotlib>=3.7.0
seaborn>=0.12.0

# Deployment
gradio>=3.30.0